rename existing DAOS files to DFS
