+++ darshan-runtime/share/ld-opts/Makefile.am
@@ -29,7 +29,7 @@ if BUILD_MDHIM_MODULE
    dist_ld_opts_DATA += darshan-mdhim-ld-opts
 endif
 if BUILD_DAOS_MODULE
-   dist_ld_opts_DATA += darshan-daos-ld-opts
+   dist_ld_opts_DATA += darshan-dfs-ld-opts
 endif
 if BUILD_APMPI_MODULE
    nodist_ld_opts_DATA += autoperf-apmpi-ld-opts
@@ -61,7 +61,7 @@ if BUILD_MDHIM_MODULE
 	echo '@$(datadir)/ld-opts/darshan-mdhim-ld-opts' >> $@
 endif
 if BUILD_DAOS_MODULE
-	echo '@$(datadir)/ld-opts/darshan-daos-ld-opts' >> $@
+	echo '@$(datadir)/ld-opts/darshan-dfs-ld-opts' >> $@
 endif
 if BUILD_APMPI_MODULE
 	echo '@$(datadir)/ld-opts/autoperf-apmpi-ld-opts' >> $@
