+++ darshan-util/darshan-dfs-logutils.c
