+++ darshan-util/darshan-logutils.h
@@ -178,7 +178,7 @@ extern struct darshan_mod_logutil_funcs *mod_logutils[];
 #include "darshan-stdio-logutils.h"
 #include "darshan-heatmap-logutils.h"
 #include "darshan-mdhim-logutils.h"
-#include "darshan-daos-logutils.h"
+#include "darshan-dfs-logutils.h"
 
 /* DXT */
 #include "darshan-dxt-logutils.h"
