+++ include/darshan-log-format.h
@@ -129,7 +129,7 @@ struct darshan_base_record
 #include "darshan-apmpi-log-format.h"
 #endif
 #include "darshan-heatmap-log-format.h"
-#include "darshan-daos-log-format.h"
+#include "darshan-dfs-log-format.h"
 
 /* X-macro for keeping module ordering consistent */
 /* NOTE: first val used to define module enum values,
@@ -177,7 +177,7 @@ struct darshan_base_record
     X(DARSHAN_APXC_MOD,     "APXC", 	  __APXC_VER,            __apxc_logutils) \
     X(DARSHAN_APMPI_MOD,    "APMPI",      __APMPI_VER,           __apmpi_logutils) \
     X(DARSHAN_HEATMAP_MOD,  "HEATMAP",    DARSHAN_HEATMAP_VER,   &heatmap_logutils) \
-    X(DARSHAN_DFS_MOD,     "DFS",        DARSHAN_DFS_VER,       &dfs_logutils)
+    X(DARSHAN_DFS_MOD,      "DFS",        DARSHAN_DFS_VER,       &dfs_logutils)
 
 /* unique identifiers to distinguish between available darshan modules */
 /* NOTES: - valid ids range from [0...DARSHAN_MAX_MODS-1]
